own class for HierarchicalMenu + Generics
HierarchicalLabel<T> can now be used with any type... not just Actions. That way HierarchicalLabel is independent of Eclipse code and could be used in other contexts too. Maybe move more to core.
